refactor: merge helpers dir into utils dir
