+++ src/components/period-picker-with-period-type/PeriodPickerWithType.test.js
@@ -11,7 +11,7 @@ import { DropDown, PeriodPicker } from '@dhis2/d2-ui-core';
 import { PeriodPickerWithPeriodType } from './PeriodPickerWithPeriodType';
 
 /* fake data */
-import fakerData from '../../helpers/fakerTests';
+import fakerData from '../../utils/fakerTests';
 
 jest.mock('@dhis2/d2-ui-core', () => ({
     DropDown: ('DropDown'),
