+++ src/pages/data-set-report/DataSetReport.js
@@ -31,8 +31,8 @@ import ShareComment from './Share';
 import Report from '../../components/report/Report';
 
 /* utils */
-import { getDocsUrl } from '../../helpers/getDocsUrl';
-import { LOADING, SUCCESS } from '../../helpers/feedbackSnackBarTypes';
+import { getDocsUrl } from '../../utils/getDocsUrl';
+import { LOADING, SUCCESS } from '../../utils/feedbackSnackBarTypes';
 
 /* styles */
 import styles from '../../styles';
