+++ src/pages/resource/add-edit-resource/AddEditResource.test.js
@@ -17,7 +17,7 @@ jest.mock('@dhis2/d2-ui-org-unit-tree', () => ({
     OrgUnitTree: ('OrgUnitTree'),
 }));
 
-import fakerData from '../../../helpers/fakerTests';
+import fakerData from '../../../utils/fakerTests';
 
 const mockResource = {
     created: "2012-11-13T13:14:26.466",
